+++ packages/atproto/lib/src/services/entities/converter/subscribed_repo_converter.dart
@@ -7,6 +7,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 
 // 🌎 Project imports:
 import '../subscribed_repo.dart';
+import '../subscribed_repo_account.dart';
 import '../subscribed_repo_commit.dart';
 import '../subscribed_repo_handle.dart';
 import '../subscribed_repo_identity.dart';
@@ -27,27 +28,31 @@ final class _SubscribedRepoConverter
 
       if (type == '#commit') {
         return SubscribedRepo.commit(
-          data: SubscribedRepoCommit.fromJson(json),
+          data: Commit.fromJson(json),
         );
       } else if (type == '#identity') {
         return SubscribedRepo.identity(
-          data: SubscribedRepoIdentity.fromJson(json),
+          data: Identity.fromJson(json),
+        );
+      } else if (type == '#account') {
+        return SubscribedRepo.account(
+          data: Account.fromJson(json),
         );
       } else if (type == '#handle') {
         return SubscribedRepo.handle(
-          data: SubscribedRepoHandle.fromJson(json),
+          data: Handle.fromJson(json),
         );
       } else if (type == '#migrate') {
         return SubscribedRepo.migrate(
-          data: SubscribedRepoMigrate.fromJson(json),
+          data: Migrate.fromJson(json),
         );
       } else if (type == '#tombstone') {
         return SubscribedRepo.tombstone(
-          data: SubscribedRepoTombstone.fromJson(json),
+          data: Tombstone.fromJson(json),
         );
       } else if (type == '#info') {
         return SubscribedRepo.info(
-          data: SubscribedRepoInfo.fromJson(json),
+          data: Info.fromJson(json),
         );
       }
 
@@ -61,6 +66,7 @@ final class _SubscribedRepoConverter
   Map<String, dynamic> toJson(SubscribedRepo object) => object.when(
         commit: (data) => data.toJson(),
         identity: (data) => data.toJson(),
+        account: (data) => data.toJson(),
         handle: (data) => data.toJson(),
         migrate: (data) => data.toJson(),
         tombstone: (data) => data.toJson(),